yachtapi/requirements.txt
azure-core==1.19.0
azure-cosmos==4.2.0
azure-storage-blob==12.8.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
Flask==1.1.2
Flask-WTF==0.15.1
idna==3.3
importlib-metadata==4.8.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
msrest==0.6.21
oauthlib==3.1.1
pycparser==2.20
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
